third_party/WebKit/LayoutTests/animations/svg-attribute-composition/svg-offset-composition-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
This is a testharness.js-based test.
PASS This test uses interpolation-test.js. 
PASS SMIL: Interpolate attribute <offset> from add [-0.5] to add [1] at (0) is [0] 
PASS SMIL: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.25) is [0.375] 
PASS SMIL: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.5) is [0.75] 
PASS SMIL: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.75) is [1.125] 
PASS SMIL: Interpolate attribute <offset> from add [-0.5] to add [1] at (1) is [1.5] 
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (-0.5) is [-0.75] assert_equals: expected "- 0.75 " but got "- 1.25 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (0) is [0] assert_equals: expected "0 " but got "- 0.5 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.25) is [0.375] assert_equals: expected "0.38 " but got "- 0.13 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.5) is [0.75] assert_equals: expected "0.75 " but got "0.25 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (0.75) is [1.125] assert_equals: expected "1.13 " but got "0.63 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (1) is [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL Web Animations: Interpolate attribute <offset> from add [-0.5] to add [1] at (1.5) is [2.25] assert_equals: expected "2.25 " but got "1.75 "
Harness: the test ran to completion.
